Split deferred_operation.h
Split the header file into three to let every class has its own header
file, also include all related changes on other files.
