+++ framework/decode/deferred_operation_info.h
@@ -83,117 +83,6 @@ class DeferredOperationInfo
     bool              process_done_;
 };
 
-class DeferredOperationInfoManager
-{
-  public:
-    static std::unique_ptr<DeferredOperationInfoManager>& Get() { return instance_; }
-
-    void add(format::HandleId deferred_operation_handle, std::unique_ptr<DeferredOperationInfo> operation)
-    {
-        if ((deferred_operation_handle != gfxrecon::format::kNullHandleId) && (operation))
-        {
-            deferred_operations_[deferred_operation_handle] = std::move(operation);
-        }
-    }
-
-    std::unique_ptr<DeferredOperationInfo>& find(format::HandleId deferred_operation_handle)
-    {
-        if (deferred_operations_.find(deferred_operation_handle) != deferred_operations_.end())
-        {
-            return deferred_operations_[deferred_operation_handle];
-        }
-        else
-        {
-            // The return of this function is a reference. If changed to nullptr
-            // , a std::unique_ptr instance will be initialized with nullptr in
-            // the function, the instance's reference will be return. But the
-            // instance will be destroyed when the function end. So caller get
-            // a reference of destroyed instance. This is the reason we use
-            // null_operation_ here.
-            return null_operation_;
-        }
-    }
-
-    void Remove(format::HandleId deferred_operation_handle)
-    {
-        std::unique_ptr<DeferredOperationInfo> deferred_operation = std::move(find(deferred_operation_handle));
-        deferred_operations_.erase(deferred_operation_handle);
-    }
-
-    DeferredOperationInfoManager() {}
-
-    ~DeferredOperationInfoManager() {}
-
-  protected:
-    std::unordered_map<format::HandleId, std::unique_ptr<DeferredOperationInfo>> deferred_operations_;
-    static std::unique_ptr<DeferredOperationInfoManager>                         instance_;
-    std::unique_ptr<DeferredOperationInfo> null_operation_ = std::unique_ptr<DeferredOperationInfo>(nullptr);
-};
-
-class DeferredOperationInfoCreateRayTracingPipelines : public DeferredOperationInfo
-{
-  public:
-    DeferredOperationInfoCreateRayTracingPipelines() :
-        DeferredOperationInfo(format::ApiCallId::ApiCall_vkCreateRayTracingPipelinesKHR)
-    {}
-
-    ~DeferredOperationInfoCreateRayTracingPipelines() {}
-
-    format::HandleId& GetDeviceId() { return device_; }
-
-    format::HandleId& GetDeferredOperationId() { return deferred_operation_; }
-
-    format::HandleId& GetPipelineCacheId() { return pipeline_cache_; }
-
-    uint32_t& GetCreateInfoCount() { return create_info_count_; }
-
-    StructPointerDecoder<Decoded_VkRayTracingPipelineCreateInfoKHR>& GetCreateInfos() { return create_infos_; }
-
-    StructPointerDecoder<Decoded_VkAllocationCallbacks>& GetAllocator() { return allocator_; }
-
-    HandlePointerDecoder<VkPipeline>& GetPipelines() { return pipelines_; }
-
-    void SetPipelineHandleInfo(std::unique_ptr<std::vector<PipelineInfo>>& pipeline_infos)
-    {
-        pipeline_handle_info_ = std::move(pipeline_infos);
-    }
-
-    std::unique_ptr<std::vector<PipelineInfo>>& GetPipelineHandleInfo() { return pipeline_handle_info_; }
-
-    void SetModifiedCreateInfos(std::unique_ptr<std::vector<VkRayTracingPipelineCreateInfoKHR>>& modified_create_infos)
-    {
-        modified_create_infos_ = std::move(modified_create_infos);
-    }
-
-    std::unique_ptr<std::vector<VkRayTracingPipelineCreateInfoKHR>>& GetModifiedCreateInfos()
-    {
-        return modified_create_infos_;
-    }
-
-    void
-    SetModifiedGroups(std::unique_ptr<std::vector<std::vector<VkRayTracingShaderGroupCreateInfoKHR>>>& modified_groups)
-    {
-        modified_groups_ = std::move(modified_groups);
-    }
-
-    std::unique_ptr<std::vector<std::vector<VkRayTracingShaderGroupCreateInfoKHR>>>& GetModifiedGroups()
-    {
-        return modified_groups_;
-    };
-
-  private:
-    format::HandleId                                                                device_;
-    format::HandleId                                                                deferred_operation_;
-    format::HandleId                                                                pipeline_cache_;
-    uint32_t                                                                        create_info_count_;
-    HandlePointerDecoder<VkPipeline>                                                pipelines_;
-    std::unique_ptr<std::vector<PipelineInfo>>                                      pipeline_handle_info_;
-    StructPointerDecoder<Decoded_VkAllocationCallbacks>                             allocator_;
-    StructPointerDecoder<Decoded_VkRayTracingPipelineCreateInfoKHR>                 create_infos_;
-    std::unique_ptr<std::vector<VkRayTracingPipelineCreateInfoKHR>>                 modified_create_infos_;
-    std::unique_ptr<std::vector<std::vector<VkRayTracingShaderGroupCreateInfoKHR>>> modified_groups_;
-};
-
 GFXRECON_END_NAMESPACE(decode)
 GFXRECON_END_NAMESPACE(gfxrecon)
