+++ framework/decode/vulkan_replay_consumer_base.cpp
@@ -30,7 +30,8 @@
 #include "decode/vulkan_enum_util.h"
 #include "decode/vulkan_feature_util.h"
 #include "decode/vulkan_object_cleanup_util.h"
-#include "decode/deferred_operation_info.h"
+#include "decode/deferred_operation_info_create_ray_tracing_pipelines.h"
+#include "decode/deferred_operation_info_manager.h"
 #include "format/format_util.h"
 #include "generated/generated_vulkan_struct_handle_mappers.h"
 #include "graphics/vulkan_device_util.h"
